module7-task3: Create useFilmLoaded hook for film and add review component
